+++ src/gameserver/states/gameplay/gamestate.ts
@@ -4,7 +4,7 @@ import { controlSystem, playerSystem } from "../../systems/coresystems";
 import { Entity } from "./entity";
 import { processClientMessages, processQueriedInputs } from "../../messaging/processclientmessages";
 import { Server } from "./../../server/server";
-import { kenneyFantasy } from "../../../../data/tilemaps/kenneyfantasy";
+import { kenneyFantasy } from "../../../modules/tilemapping/tilemaps/kenneyfantasy";
 import { Mesh } from "three";
 
 /**
